backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
pdfplumber>=0.10
python-dotenv>=1.0
groq>=0.5
